+++ first-service/src/main/java/itmo/tuchin/nikitin/first_service/dto/CoordinatesUpdateDTO.java
@@ -12,6 +12,6 @@ public class CoordinatesUpdateDTO {
     private String x; //Максимальное значение поля: 682, Поле не может быть null
 
     @Min(-984)
-    @Max(Integer.MAX_VALUE)
+    @Max(Long.MAX_VALUE)
     private String y; //Значение поля должно быть больше -985
 }
